Implement FS#3 - Method for building multiple targets at once
Multiple targets can now be specified as multiple positional arguments
on the commandline. (i.e., ./make.py debug release). Additionally, there
is now an option for --all-targets that will build all targets that are
specified anywhere in the makefile.

This revision also improves some of the help text formatting, and fixes
an issue with formatting of the "Compile of <project> took <time>"
messages.